refactor: 입력받은 다리 길이의 검증 로직을 또 다른 클래스로 분리
